+++ src/scip/projection/__init__.py
@@ -1,0 +1,3 @@
+
+def project_block_partition(part, proj, **proj_kw):
+    return [proj(p, **proj_kw) for p in part]
