+++ test/features/test_feature_extraction.py
@@ -1,11 +1,11 @@
-from scip.features import feature_extraction
+from scip.features import extract_features
 from dask.dataframe import DataFrame
 import pytest
 
 
 @pytest.mark.parametrize("fake_images_bag", [True, False], indirect=True)
 def test_extract_features(fake_images_bag, fake_image_nchannels):
-    features = feature_extraction.extract_features(
+    features = extract_features(
         images=fake_images_bag,
         channel_names=[f"c_{i}" for i in range(fake_image_nchannels)],
         types=["bbox", "intensity", "shape", "texture"]
